Tolak akses non-API menuju Post

+++ application/controllers/api.php
@@ -83,44 +83,47 @@ class Api extends CI_Controller {
 
 	public function post() {
 		header("Access-Control-Allow-Origin: *");
+		if ($this->input->post('katalog') && $this->input->post('jenis') && $this->input->post('warna')) {
+			$this->db->truncate('katalog');
+			foreach($this->input->post('katalog') as $katalog) {
+				$data = array(
+					'rowid' => $katalog['rowid'],
+					'nama' => $katalog['nama'],
+					'jenis' =>$katalog['jenis'],
+					'warna' => $katalog['warna'],
+					'keterangan' => $katalog['keterangan'],
+					'dicuci' => $katalog['dicuci'],
+					'timeselesai' => $katalog['timeselesai']
+				);
+				$this->db->insert('katalog', $data);
+			}
 
-		$this->db->truncate('katalog');
-		foreach($this->input->post('katalog') as $katalog) {
-			$data = array(
-				'rowid' => $katalog['rowid'],
-				'nama' => $katalog['nama'],
-				'jenis' =>$katalog['jenis'],
-				'warna' => $katalog['warna'],
-				'keterangan' => $katalog['keterangan'],
-				'dicuci' => $katalog['dicuci'],
-				'timeselesai' => $katalog['timeselesai']
-			);
-			$this->db->insert('katalog', $data);
+			$this->db->truncate('jenis');
+			foreach($this->input->post('jenis') as $jenis) {
+				$data = array(
+					'rowid' => $jenis['rowid'],
+					'nama' => $jenis['nama'],
+					'jenis' =>$jenis['jenis'],
+					'jumlah' => $jenis['jumlah']
+				);
+				$this->db->insert('jenis', $data);
+			}
+
+			$this->db->truncate('warna');
+			foreach($this->input->post('warna') as $warna) {
+				$data = array(
+					'rowid' => $warna['rowid'],
+					'warna' => $warna['warna'],
+					'hex' => $warna['hex']
+				);
+				$this->db->insert('warna', $data);
+			}
+
+
+			$this->db->where('rowid', 1)->update('timestamp', array('timestamp' => $this->input->post('timestamp')));
 		}
 
-		$this->db->truncate('jenis');
-		foreach($this->input->post('jenis') as $jenis) {
-			$data = array(
-				'rowid' => $jenis['rowid'],
-				'nama' => $jenis['nama'],
-				'jenis' =>$jenis['jenis'],
-				'jumlah' => $jenis['jumlah']
-			);
-			$this->db->insert('jenis', $data);
-		}
-
-		$this->db->truncate('warna');
-		foreach($this->input->post('warna') as $warna) {
-			$data = array(
-				'rowid' => $warna['rowid'],
-				'warna' => $warna['warna'],
-				'hex' => $warna['hex']
-			);
-			$this->db->insert('warna', $data);
-		}
-
-
-		$this->db->where('rowid', 1)->update('timestamp', array('timestamp' => $this->input->post('timestamp')));
+		else echo "Akses ditolak.";
 	}
